rm character/builder
